[#43, #45] DBService squashed commits
Force the API service to wait for the DB service to be running before starting the API.

Modify entry point to wait for the DB service to spin up first.

Correct the image name with the dev tag.

Modify the application's port to match the port stated in README.md

Example call to the DB service.

Change application port to match the one stated in README.md